intermediate merge

+++ VinezorsiOS/VinezorsiOS/History.h
@@ -19,7 +19,7 @@ enum CoverMode { FIRST_TIME, PLAYING };
 struct History
 {
     Ogre::SceneManager* sceneMgr;
-
+    
     int nback;
     std::vector<SceneNode*> seats;
     std::vector<Pod*> list;
@@ -31,16 +31,16 @@ struct History
     void initSeats(int nback);
     void addPod(const PodInfo & podInfo);
     void revealPod();
-    void update(double elapsed);
+    void update(float elapsed);
     void setCoverDest(int seatInd);
     void determineCoverLoc(bool success);
     
-    double panelHeight;
-    double panelX;
-    double panelY;
-    double reservedHeight;
-    double coverHeight;
-    double coverOffsetY;
+    float panelHeight;
+    float panelX;
+    float panelY;
+    float reservedHeight;
+    float coverHeight;
+    float coverOffsetY;
     
     int saveInd;
     int coverInd;
